Use include_bytes! instead of incbin.
This is the second attempt at this patch. The first version was reverted
in b22f1d688e80c664775ec29db088b9aef4fffaed

The problem, I suspect, was that the snapshot was represented as a
source_set, which inserted a node into the dependency tree.
include_bytes does properly insert the snapshot into rustc's depfile but
the use of source_set confused gn. Now the that the deno executable has
the create_deno_snapshot as a direct dependency, changes will be
propagated.